Add command executor class for compiling and testing repo
