--- app/src/test/java/ciserver/PipelinePullerTest.java
@@ -1,42 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package ciserver;
-
-import static org.junit.Assert.*;
-
-import java.io.File;
-import java.io.IOException;
-
-import org.apache.commons.io.FileUtils;
-import org.eclipse.jgit.api.Git;
-import org.junit.Test;
-
-import com.google.gson.Gson;
-
-public class PipelinePullerTest {
-
-    @Test
-    public void canPull() throws IOException {
-        var gson = new Gson();
-        var event = gson.fromJson(PushEventTest.TestData, PushEvent.class);
-        var pipeline = new Pipeline(event, PipelineTest.PipelineTestingDirectory);
-
-        var status = pipeline.start(Target.PULL);
-        assertEquals(status, PipelineStatus.Ok);
-
-        File directory = new File(
-                String.format("%s/%s/%s", PipelineTest.PipelineTestingDirectory, event.repository.name,
-                        event.headCommit.id));
-        assertTrue(directory.isDirectory());
-
-        try {
-            var repo = Git.open(directory);
-            assertEquals(event.headCommit.id, repo.getRepository().getBranch());
-        } catch (IOException e) {
-            assertTrue("Could not open git project", false);
-        } finally {
-            FileUtils.deleteDirectory(directory);
-        }
-    }
-}
